+++ src/proxy/AST/AST.js
@@ -6,13 +6,15 @@
 /**
  * Parse rewrite source file to AST tree
  * @param sourceCode
+ * @param filePath
  * @returns {{}}
  */
-module.exports = function parseRewrite(sourceCode){
+module.exports = function parseRewrite(sourceCode, filePath){
     var res = {
         baseRules: [],
         domains: [],
-        commands: []
+        commands: [],
+        filePath: filePath
     };
     var target = res;
     var history = [];
@@ -39,7 +41,7 @@ module.exports = function parseRewrite(sourceCode){
         baseRule: /^(.*?\s*=>\s*[^\{\}]*)$/,
         cmd: /^(\w+(?:\s[^\{]+)+)$/,
         // rule: /(.*?\s*=>\s*\{[\s\S]*?\})/,
-        domainStart: /^([^\/]+) => \{$/,
+        domainStart: /^(([^\/]+) => \{)|(domain ([^\/]+) \{)$/,
         locationStart: /^location\s((~\s*)?\/.*?)+\s*\{$/,
         end: /^}$/
     };
@@ -66,8 +68,12 @@ module.exports = function parseRewrite(sourceCode){
                         break;
 
                     case 'domainStart':
+                        var domain = line.indexOf('domain') === 0
+                            ? line.replace(/domain ([^\/]+) \{/, '$1')
+                            : line.split(/\s*=>\s*/)[0];
+
                         target.domains.push({
-                            domain: line.split(/\s*=>\s*/)[0],
+                            domain: domain,
                             commands: [],
                             location: [],
                             props: {}