+++ src/proxy/commands/setHeader.js
@@ -4,11 +4,14 @@
  */
 'use strict';
 
-module.exports = function setHeader(response, key, val){
-    if(typeof response.setHeader === 'function'){
-        response.setHeader(key, val);
-    }else{
-        response.headers = response.headers || {};
-        response.headers[key] = val;
+module.exports = function setHeader(response, name, value){
+    var header = response.getHeader(name);
+
+    if(!Array.isArray(header)){
+        header = !!header ? [header] : [];
     }
+
+    header.push(value);
+
+    response.setHeader(name, header);
 };