+++ src/proxy/server/Server.js
@@ -2,8 +2,10 @@
  * @file
  * @author zdying
  */
+var https = require('https');
 var http = require('http');
 var path = require('path');
+var tls = require('tls');
 var url = require('url');
 var net = require('net');
 var fs = require('fs');
@@ -12,6 +14,7 @@ var aliasWorker = require('./workers/alias');
 var requestWorker = require('./workers/request');
 
 var merge = require('../../helpers/merge');
+var config = require('../../client/config');
 
 var parseHosts = require('../tools/parseHosts');
 var parseRewrite = require('../tools/parseRewrite');
@@ -24,7 +27,10 @@ var logger = log.namespace('proxy -> Server');
 
 //TODO 支持rewrite到hosts中的host时
 
-function Server(){}
+function Server(port, middleManPort){
+    this.port = port || 4936;
+    this.middleManPort = middleManPort || 10010;
+}
 
 Server.cache = {
     hosts: {},
@@ -44,15 +50,50 @@ Server.prototype = {
         this.domainCache = {};
     },
 
-    start: function(port, option){
+    start: function(option){
         this.init();
-
-        this.port = Number(port) || 4936;
 
         this.server = http.createServer()
             .listen(this.port);
 
+        this.middleManServer = this.createMiddleManServer()
+            .listen(this.middleManPort);
+
         return new Promise(this.initEvent.bind(this));
+    },
+
+    createMiddleManServer: function(){
+        // HTTPS Middle Man Server
+        var defaultCert = {
+            key: path.resolve(__dirname, '../../../ssl/cert/localhost.key'),
+            cert: path.resolve(__dirname, '../../../ssl/cert/localhost.crt')
+        };
+
+        var option = {
+            key: fs.readFileSync(defaultCert.key),
+            cert: fs.readFileSync(defaultCert.cert),
+            SNICallback: function (domain, cb) {
+                var domainCache = this.domainCache;
+                var domainRewriteRule = domainCache[domain] || [];
+                var certObj = domainRewriteRule.length > 0 && domainRewriteRule[0].props;
+
+                if(certObj && certObj.sslCertificateKey && certObj.sslCertificate) {
+                    cb(null, tls.createSecureContext({
+                        key: fs.readFileSync(certObj.sslCertificateKey),
+                        cert: fs.readFileSync(certObj.sslCertificate)
+                    }));
+                    log.debug('SNI callback [', domain.bold.green, ']:', JSON.stringify(certObj));
+                } else {
+                    cb(null, tls.createSecureContext({
+                        key: fs.readFileSync(defaultCert.key),
+                        cert: fs.readFileSync(defaultCert.cert)
+                    }));
+                    log.warn('No keys/certificates for domain requested:', domain.bold.yellow);
+                }
+            }.bind(this)
+        };
+
+        return https.createServer(option)
     },
 
     initEvent: function(resolve, reject){
@@ -60,6 +101,7 @@ Server.prototype = {
         var url = 'http://127.0.0.1:' + port;
         var pac = url + '/proxy.pac';
         var server = this.server;
+        var middleManServer = this.middleManServer;
 
         server
             .on('listening', function(){
@@ -76,7 +118,35 @@ Server.prototype = {
                 reject(err)
             })
             .on('request', this.requestHandler.bind(this))
-            .on('connect', this.connectHandler.bind(this))
+            .on('connect', this.connectHandler.bind(this));
+
+        middleManServer
+            .on('request', function(req, res){
+                var url = req.url;
+                var host = req.headers.host;
+                var protocol = req.client.encrypted ? 'https' : 'http';
+
+                log.debug('http middle man server receive request ==>', protocol, host, url);
+
+                if(!url.match(/^\w+:\/\//)){
+                    req.url = protocol + '://' + host + url;
+                }
+
+                if(host === '127.0.0.1:' + this.middleManPort){
+                    res.end('the man in the middle page: ' + url);
+                    // if(url === '/'){
+                    //     res.end('the man in the middle.');
+                    // }else if(url === '/favicon.ico'){
+                    //     res.statusCode = 404;
+                    //     res.end('404 Not Found.');
+                    // }else{
+                    //     res.statusCode = 404;
+                    //     res.end('404 Not Found.');
+                    // }
+                }else{
+                    this.requestHandler(req, res);
+                }
+            }.bind(this));
     },
 
     stop: function(){
@@ -226,13 +296,15 @@ Server.prototype = {
 
         this.updateDomainCache();
 
-        logger.debug('rewriteRules updated =>', JSON.stringify(this.rewriteRules));
+        logger.debug('rewriteRules updated ==>', (this.rewriteRules || []).map(function(rule){
+            return Object.keys(rule.domains)
+        }));
+        logger.detail('rewriteRules updated ==>', JSON.stringify(this.rewriteRules));
     },
 
     requestHandler: function(request, response){
         var _url = request.url;
         var start = Date.now();
-        var pacFilePath = path.resolve(__hii__.cacheTmpdir, 'hiipack.pac');
 
         if(_url === '/'){
             response.end('proxy file url: http://127.0.0.1:' + this.port + '/proxy.pac');
@@ -240,8 +312,10 @@ Server.prototype = {
         }
 
         if(_url === '/proxy.pac'){
+            var pacFilePath = path.resolve(__hii__.cacheTmpdir, 'hiipack.pac');
+
             fs.readFile(pacFilePath, 'utf-8', function(err, str){
-                 response.end(str);
+                response.end(str);
             });
             return
         }
@@ -265,6 +339,7 @@ Server.prototype = {
     connectHandler: function(request, socket, head){
         var urlObj = url.parse('https://' + request.url);
         var domains = this.domainCache[urlObj.hostname];
+        var middleManPort = this.middleManPort;
 
         if(Array.isArray(domains)){
             domains.forEach(function(domain){
@@ -274,25 +349,31 @@ Server.prototype = {
                     proxy = proxy.location[0].props.proxy;
 
                     urlObj = url.parse(proxy);
-                    urlObj.port = 443;
+                    urlObj.hostname = '127.0.0.1';
+                    urlObj.port = middleManPort;
 
                     logger.info('https proxy -', request.url.bold.green, '==>', urlObj.hostname.bold.green);
                 }else if(typeof domain === 'string'){
                     // hosts规则
                     urlObj = url.parse('https://' + domain);
-                    urlObj.port = 443;
+                    urlObj.hostname = '127.0.0.1';
+                    urlObj.port = middleManPort;
                 }else{
                     logger.info('https direc -', request.url.bold);
                 }
             });
+        }else{
+            logger.info('https direc -', request.url.bold);
         }
+
+        log.debug('connect to:', urlObj.port, urlObj.hostname);
 
         var proxySocket = net.connect(urlObj.port, urlObj.hostname, function(){
             socket.write('HTTP/1.1 200 Connection Established\r\n\r\n');
             proxySocket.write(head);
             proxySocket.pipe(socket);
         }).on('error', function(e){
-            console.log('proxy error', e.message);
+            logger.error('proxy error', e.message);
             socket.end();
         }).on('data', function(data){
             // console.log('proxy socker data:', data.toString());
@@ -354,10 +435,10 @@ Server.prototype = {
             }
         });
 
-        createPacFile(domainCache);
+        createPacFile(this.port, domainCache);
 
-        logger.debug('domain cache updated', Object.keys(domainCache));
-        logger.detail('domain cache updated', JSON.stringify(domainCache));
+        logger.debug('domain cache updated ==>', Object.keys(domainCache));
+        logger.detail('domain cache updated ==>', JSON.stringify(domainCache));
     }
 };