+++ src/proxy/server/workers/request.js
@@ -5,36 +5,83 @@
 
 'use strict';
 
-var url = require('url');
-var http = require('http');
-
 var logger = log.namespace('proxy -> Server');
 
 var execResponseCommand = require('../execCommand');
 
 module.exports = {
     response: function(rewrite_rule, request, response){
-        var _url = request.url;
-        var uri = url.parse(_url);
+        var url = require('url');
+        var http = require('http');
+        var https = require('https');
+        var zlib = require('zlib');
 
-        var proxy = http.request(request.proxy_options, function(res){
+        var proxyOption = request.proxy_options;
+        var isHTTPS = proxyOption.protocol === 'https:';
+
+        proxyOption.headers['accept-encoding'] = 'gzip,deflate';
+
+        if(isHTTPS){
+            // proxyOption.port = 443;
+            proxyOption.rejectUnauthorized = false;
+        }
+
+        var proxy = (isHTTPS ? https : http).request(proxyOption, function(res){
+            response.headers = res.headers;
+
+            var encoding = response.headers['content-encoding'];
+            // response.removeHeader('Content-Encoding');
+            // delete response.headers['content-encoding'];
+
             execResponseCommand(rewrite_rule, {
-                response: res
+                response: response
             }, 'response');
 
             // response.pipe(res);
             response.writeHead(res.statusCode, res.headers);
-            res.on('data', function(chunk){
-                response.write(chunk);
-            });
-            res.on('end', function(){
-                var proxyOption = request.proxy_options;
 
+            /*
+            res.pause();
+
+            log.warn('request was paused:'.red, _url.bold.red);
+
+            setTimeout(function(){
+                res.resume();                
+                log.warn('request was resumed:'.red, _url.bold.red);
+            }, 5000)
+            */
+
+            var contentType = res.headers['content-type'];
+
+            if(global.program && global.program.proxyContentLog && /(text|json|javascript|css)/.test(contentType)){
+                // 打印日志
+                if(encoding === 'gzip' || encoding === 'deflate'){
+                    var unzipStream = zlib.createUnzip();
+
+                    unzipStream.on('data', function(chunk){
+                        console.log('ondata =>', chunk.toString());
+                    });
+
+                    unzipStream.on('error', function(err){
+                        console.log('error ==>', err);
+                    });
+
+                    res.pipe(unzipStream);
+                }else{
+                    res.on('data', function(chunk){
+                        console.log('ondata =>', chunk.toString());
+                    })
+                }
+            }
+
+            res.pipe(response);
+
+            res.on('end', function(){
                 request.res = res;
-                response.end();
 
                 if(request.PROXY){
-                    logger.access(request, (uri.protocol || 'http:') + '//' + proxyOption.host + (proxyOption.port ? ':' + proxyOption.port : '') + proxyOption.path)
+                    logger.access(request, (proxyOption.protocol || 'http:') + '//' + proxyOption.hostname +
+                        (proxyOption.port ? ':' + proxyOption.port : '') + proxyOption.path)
                 }else{
                     logger.access(request);
                     // logger.info('direc -', request.url.bold, Date.now() - start, 'ms');