+++ src/proxy/tools/createPacFile.js
@@ -7,9 +7,13 @@
 
 var fs = require('fs');
 var path = require('path');
-var config = require('../../client/config');
+var os = require('os');
+var dns = require('dns');
 
-module.exports = function createPacFile(domainsCache){
+var config = require('../../client/config');
+var localIP = '127.0.0.1';
+
+module.exports = function createPacFile(proxyPort, domainsCache){
     // 这个方法是生成到proxy.pac中的，hiipack不会调用
     function FindProxyForURL(url, host) {
         host = host.toLowerCase();
@@ -54,19 +58,30 @@ module.exports = function createPacFile(domainsCache){
         }
     };
 
-    var txt = [
-        'var SYS_PROXY = "' + (sysProxy ? 'PROXY ' + sysProxy : '') + '";\n',
-        'var PROXY = "PROXY 127.0.0.1:4936";\n',
-        'var DIRECT = "DIRECT";\n',
-        'var EXCLUDE_REG = /' + regText + '/;\n',
-        'var DOMAINS = ' + JSON.stringify(domainsCache, replaceFun, 4) + ';\n\n',
+    // 获取本机IP
+    dns.resolve(os.hostname(), function(err, addr){
+        if(err){
+            log.warn('Get local ip failed, use 127.0.0.1');
+            log.warn(err.message);
+        }else{
+            localIP = Array.isArray(addr) ? addr[0] : addr;
+            log.debug('localIP is ==>', addr);
+        }
 
-        FindProxyForURL.toString().replace(/^\s{8}/mg, '')
-    ];
+        var txt = [
+            'var SYS_PROXY = "' + (sysProxy ? 'PROXY ' + sysProxy : '') + '";\n',
+            'var PROXY = "PROXY ' + localIP + ':' + proxyPort+ '";\n',
+            'var DIRECT = "DIRECT";\n',
+            'var EXCLUDE_REG = /' + regText + '/;\n',
+            'var DOMAINS = ' + JSON.stringify(domainsCache, replaceFun, 4) + ';\n\n',
 
-    var pacFilePath = path.resolve(__hii__.cacheTmpdir, 'hiipack.pac');
+            FindProxyForURL.toString().replace(/^\s{8}/mg, '')
+        ];
 
-    fs.writeFile(pacFilePath, txt.join(''), function(err){
-        err && log.error(err);
+        var pacFilePath = path.resolve(__hii__.cacheTmpdir, 'hiipack.pac');
+
+        fs.writeFile(pacFilePath, txt.join(''), function(err){
+            err && log.error(err);
+        });
     });
 };